requirements.txt
fastapi
uvicorn
pydantic
pyswisseph
langchain
langchain-core
langchain-openai
openai
pytz
python-dotenv
supabase
tiktoken
anyio
